RAG_Projects/Social Media App Support Agent/requirements.txt
langchain==0.3.22
langchain-core==0.3.49
langchain-community==0.3.20
langchain-text-splitters==0.3.7
faiss-cpu==1.10.0
openai==1.70.0
langchain-openai==0.3.11
python-dotenv==1.0.1
pytest==8.3.4
pytest-cov==6.0.0
streamlit==1.36.0